add web flavor of mdx
