--- apps/demo/src/supported/table-html.mdx
@@ -1,20 +1,0 @@
-<table>
-  <tr>
-    <th>Header 1</th>
-    <th colSpan="2">Header 2 & 3</th>
-  </tr>
-  <tr>
-    <td>Row 1, Cell 1</td>
-    <td>Row 1, Cell 2</td>
-    <td>Row 1, Cell 3</td>
-  </tr>
-  <tr>
-    <td rowSpan="2">Row 2 & 3, Cell 1</td>
-    <td>Row 2, Cell 2</td>
-    <td>Row 2, Cell 3</td>
-  </tr>
-  <tr>
-    <td>Row 3, Cell 2</td>
-    <td>Row 3, Cell 3</td>
-  </tr>
-</table>
