+++ package.json
@@ -20,7 +20,7 @@
     "prettier": "^2.2.1",
     "ts-node": "^10.0.0",
     "typescript": "^4.2.4",
-    "expo-module-scripts": "^2.0.0",
+    "expo-module-scripts": "^3.5.2",
     "@types/jest": "^26",
     "jest": "^26.6.3"
   }
