+++ packages/mdx/src/index.tsx
@@ -1,4 +1,6 @@
 export { MDXStyles, useMDXStyles } from "./MDXStyles";
+export { getDOMComponents } from "./getDOMComponents";
+export { getUniversalComponents } from "./getUniversalComponents";
 export { MDXComponents } from "./MDXComponents";
 export { useMDXComponents } from "./useMDXComponents";
 export { resolveAssetUrl } from "./resolveAssetUrl";
